camera: add manual camera control to TR2
This allows moving the camera manually in TR2, similar to TR1, with
WASD. The logic is largely shifted to TRX from TR1, with Camera_Update
in both games to follow separately.

Resolves #2978.
