+++ docs/tr2/README.md
@@ -283,6 +283,7 @@ However, you can easily download them manually from these urls:
 - added ability to hold arrows to move through menus more quickly
 - added additional custom control schemes
 - added customizable controller support
+- added the ability to move the camera around with W,A,S,D (rebindable)
 - fixed setting user keys being very difficult
 - fixed skipping FMVs triggering inventory
 - fixed skipping credits working too fast
