+++ src/libtrx/config/map_tr2.def
@@ -74,5 +74,6 @@ CFG_INT32(g_Config, audio.music_volume, 10)
 CFG_BOOL(g_Config, audio.enable_lara_mic, false)
 CFG_ENUM(g_Config, audio.underwater_music_mode, UMM_FULL, UNDERWATER_MUSIC_MODE)
 CFG_ENUM(g_Config, audio.music_load_condition, MUSIC_LOAD_NON_AMBIENT, MUSIC_LOAD_CONDITION)
+CFG_INT32(g_Config, gameplay.camera_speed, 5)
 CFG_BOOL(g_Config, gameplay.enable_game_modes, true)
 CFG_BOOL(g_Config, profile.new_game_plus_unlock, false)
