+++ src/libtrx/config/priv_tr2.c
@@ -155,6 +155,7 @@ void Config_Sanitize(void)
         g_Config.rendering.fps = 30;
     }
 
+    CLAMP(g_Config.gameplay.camera_speed, 1, 10);
     CLAMP(g_Config.visuals.fog_start, 1, 100);
     CLAMP(g_Config.visuals.fog_end, 1, 100);
     CLAMP(g_Config.visuals.fov, 30, 150);
