+++ src/libtrx/include/libtrx/config/types_tr2.h
@@ -100,6 +100,7 @@ typedef struct {
         bool enable_cutscenes;
         bool enable_auto_item_selection;
         int32_t turbo_speed;
+        int32_t camera_speed;
         bool enable_game_modes;
     } gameplay;
