+++ src/tr1/game/camera/common.h
@@ -5,4 +5,3 @@
 #include <stdint.h>
 
 void Camera_UpdateCutscene(void);
-void Camera_MoveManual(void);
